Use a ReverseProxy for HTTP intercepts.
Forwarding HTTP requests by reading and writing the request in full
using `http.ReadRequest` and `request.Write` works, but it has some
severe implications.

1. An internal buffer is required that must be large enough to contain
   the entire request. This isn't great when the request is streaming
   lots of data.
2. The `http.ReadRequest` only supports HTTP/1.x requests. HTTP/2.0 is
   very common today. GRPC, for instance, is using HTTP/2.0.
3. Further enhancements like adding TLS or websocket support would be
   much easier if Telepresence could use a proper HTTP server.

This commit replaces the current logic with a single-host `http.Server`
that dispatches request using a `httputil.ReverseProxy`. The proxy
will either dial the application container or use a dialer that returns
a connection that is based on `tunnel.Stream` to the client.

A `net.Listener` is wrapped in a `ListenerSwitch` which provides the
primary and secondary listeners that ultimately will call `Accept`. The
secondary listener is used whenever a HTTP-filtered intercept or wiretap
is active.

Signed-off-by: Thomas Hallgren <thomas@tada.se>