Give the Telepresence Daemons better names
Using the name xxx-foreground isn't very intuitive when talking about
daemon processes. Yes, the command, when issued in a terminal, will
start the daemon in foreground so the names of the commands does have
some logic to them, but then again, starting in the foreground is the
default behavior of any command. And when the same command is started
from the CLI, it will be started in the background, despite its name.

The daemons are now renamed:

- connector-foreground => userd
- daemon-foreground => rootd
- kubeauth-foreground => kubeauthd

Signed-off-by: Thomas Hallgren <thomas@tada.se>