Remove unnecessary setcap from traffic binary
The setcap capability was added to allow the traffic binary to bind to privileged ports (specifically port 443 for the mutating webhook). Since version 2.24.0, the default mutating webhook port was changed to 8443 (a non-privileged port), making this capability unnecessary. This change removes the setcap command from the Dockerfile, simplifying the build process and reducing security surface area.

Signed-off-by: Błażej Gruszka <blazej.gruszka@displate.com>