Error on connect with --docker in combination with intercept.useFtp=true
The `--docker` flag is not allowed when the client is configured with
`intercept.useFtp=true` and an error is now generated instantly by the
`telepresence connect --docker` command.

The docker volume plugin cannot use FTP because it requires two ports:
a fixed control port that Telepresence can proxy, and a dynamic data
port (randomly chosen during connection) that Telepresence cannot proxy
on-demand. The port-forwarder only forwards pre-configured ports and
doesn't understand FTP's protocol. Consequently, FTP isn't allowed in
this scenario. If it was, then when an FTP server tells the client to
use an unpredictable second port for file transfers, Telepresence would
block it—causing the connection to fail every time.

Signed-off-by: Thomas Hallgren <thomas@tada.se>