Refactor pod resolution to handle multiple pods in `portforward`.
This commit replaces the `getFirstPod` function with logic that
processes a list of pods, selecting the first running pod with a valid
container port.

The fix addresses an intermittent failure in `Test_HelmTemplateInstall`
because it installs another pod into the traffic-manager namespace that
uses the same labels as the traffic-manager pod.

Signed-off-by: Thomas Hallgren <thomas@tada.se>