Always pass kubeconfigs as kubeconfig-data
Switch from writing the modified kubeconfig to a file in the user's
home directory to passing its contents directly in the `kubeconfig-data`
field.

This removes the need for the daemon to access any home directories,
enabling the systemd unit to use the stronger `ProtectHome=yes` (fully
inaccessible) rather than `ProtectHome=read-only`.

Signed-off-by: Thomas Hallgren <thomas@tada.se>