Make traffic-agent consumption metrics reporting optional.
Metrics reporting for the traffic-agent can now be optionally enabled or
disabled via the `agent.enableConsumptionMetrics` Helm chart value.
The traffic-agent metrics will also be completely disabled when the
Helm chart value `prometheus.port` is unset or zero.

This change is intended to reduce the amount of unnecessary traffic
generated by the traffic-agent when consumption metrics reporting is of
no interest to the user.

Signed-off-by: Thomas Hallgren <thomas@tada.se>